// apps/b2c/src/shared/types/app-header-copy.ts
export type AppHeaderCopy = {
  language: string
  languageDropdownTitle: string
  currency: string
  currencyModalTitle: string
  currencyFeaturedSectionTitle: string
  currencyAllSectionTitle: string
  notificationTitle: string
  notificationEmptyLabel: string
  authLabel: string
  navFlights: string
  navHotels: string
  navTransfers: string
  navExcursions: string
  navCarRental: string
  navTourPackages: string
  navAdditionalServices: string
}